Addding error handling for writings outside of transaction.
