+++ src/common/object/node/collection.hpp
@@ -16,14 +16,10 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
-#if REALM_PLATFORM_NODE
-#include "node/subscriber.hpp"
-#include "node/methods.hpp"
-#include "node/collection.hpp"
-#include "node/object.hpp"
-#else
-#include "jsc/subscriber.hpp"
-#include "jsc/methods.hpp"
-#include "jsc/collection.hpp"
-#include "jsc/object.hpp"
-#endif
+#pragma once
+
+struct IOCollection {
+    virtual void set(Napi::Env, std::string, Napi::Value) = 0;
+    virtual Napi::Value get(Napi::Env, std::string) = 0;
+};
+
