+++ src/common/object/observer.hpp
@@ -16,14 +16,10 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
-#if REALM_PLATFORM_NODE
-#include "node/subscriber.hpp"
-#include "node/methods.hpp"
-#include "node/collection.hpp"
-#include "node/object.hpp"
-#else
-#include "jsc/subscriber.hpp"
-#include "jsc/methods.hpp"
-#include "jsc/collection.hpp"
-#include "jsc/object.hpp"
-#endif
+#pragma once
+
+struct ObjectObserver {
+    virtual void subscribe(Subscriber *) = 0;
+    virtual void remove_subscription(const Subscriber *) = 0;
+    virtual void unsubscribe_all() = 0;
+};
