+++ src/test/test.sh
@@ -1,4 +1,3 @@
-clear &&
 g++ -g -std=c++17 -Wall -I catch.hpp -I test_bed.hpp -I ../ \
     -I ../../vendor/realm-core/src/ \
     -I ../..//cmake-build-debug/vendor/realm-core/src/ \
